+++ Rebus.Tests/Synchronous/TestSyncBusNoBlocking.cs
@@ -90,12 +90,12 @@ async implementation underneath the covers without deadlocking, even in a single
                 {
                     using (var context = new DefaultSyncTransactionContext())
                     {
-                        AmbientTransactionContext.Current = context;
+                        var transactionContext = AmbientTransactionContext.Current;
 
                         try
                         {
                             // enlist some other async thing
-                            context.OnCommitted(async () =>
+                            transactionContext.OnCommitted(async () =>
                             {
                                 Console.WriteLine("waiting....");
                                 await Task.Delay(100);
@@ -112,7 +112,7 @@ async implementation underneath the covers without deadlocking, even in a single
                         }
                         finally
                         {
-                            AmbientTransactionContext.Current = null;
+                            AmbientTransactionContext.SetCurrent(null);
                         }
                     }
                 }, null);