+++ Rebus/Pipeline/Receive/HandleDeferredMessagesStep.cs
@@ -87,7 +87,7 @@ This is done by checking if the incoming message has a '" + Headers.DeferredUnti
                         transportMessage.Headers[Headers.MessageId],
                         returnAddress);
 
-                    using (var context = new DefaultTransactionContext())
+                    using (var context = new TransactionContext())
                     {
                         await _transport.Send(returnAddress, transportMessage, context);